refactor: reuse ShareScreen component
